Implemented table update on date change
