+++ src/Header.js
@@ -1,23 +1,16 @@
-import React, { useState } from 'react';
+import React from 'react';
 import { hot } from 'react-hot-loader';
 import Dates from './Dates';
 import "./Header.css";
 
 function Header(props) {
-  const [date, setDate] = useState(new Date());
-
-  function handleDateChange(event) {
-    var date = new Date(event.target.value);
-    setDate(date);
-  }
-
   return (
     <div className="header">
       <div className="data-input">
         <span>From: </span>
-        <input type="date" id="fromDate" defaultValue={date.toISOString().substr(0,10)} onChange={handleDateChange}/>
+        <input type="date" id="fromDate" defaultValue={props.date.toISOString().substr(0,10)} onChange={props.onDateChange}/>
       </div>
-      <Dates date={date} columns={props.columns}/>
+      <Dates date={props.date} columns={props.columns}/>
     </div>
   );
 }
