+++ src/app.js
@@ -1,4 +1,4 @@
-import React, { useLayoutEffect, useState } from "react";
+import React, { useLayoutEffect, useState, useReducer } from "react";
 import { hot } from "react-hot-loader";
 import Table from "./Table";
 import Header from "./Header";
@@ -11,10 +11,139 @@ function App(props) {
   var containerWidth = remToPx(4);
   var columns = Math.ceil((width - roomCellWidth) / containerWidth);
 
+  const [date, setDate] = useState(new Date());
+
+  const rooms = {
+    "floors": [
+      {
+        "name": "piano 1",
+        "rooms": [
+          {
+            "number": 1,
+            "type": "camera matrimoniale/doppia"
+          },
+          {
+            "number": 2,
+            "type": "appartamento"
+          },
+          {
+            "number": 3,
+            "type": "camera tripla"
+          }
+        ]
+      },
+      {
+        "name": "piano 2",
+        "rooms": [
+          {
+            "number": 6,
+            "type": "camera matrimoniale/doppia"
+          },
+          {
+            "number": 7,
+            "type": "camera matrimoniale/doppia"
+          },
+          {
+            "number": 8,
+            "type": "camera singola"
+          }
+        ]
+      }
+    ]
+  };
+
+  const tiles = [
+    {
+      name: "Ivan Petrov",
+      colour: "rgba(217, 73, 73, 0.69)",
+      roomType: "doppia",
+      roomNumber: 2,
+      from: "20210729",
+      nights: 2
+    }
+  ];
+
+  function handleMove(occupations, x, y, pageY) {
+    var margin = remToPx(8) + 1;
+    var tableY = pageY - margin;
+    var rowHeight = remToPx(4) + 1;
+    var targetRow = Math.floor(tableY / rowHeight);
+    var targetY = -1;
+    for (var i = 0; i < rooms.floors.length; i++) {
+      const floor = rooms.floors[i];
+      if (targetRow == 0) {
+        break;
+      }
+      targetRow--;
+
+      for (var j = 0; j < floor.rooms.length; j++) {
+        const room = floor.rooms[j];
+        if (targetRow == 0) {
+          targetY = room.number;
+          break;
+        }
+        targetRow--;
+      }
+    }
+
+    if (targetY > 0) {
+      var newOccupations = [...occupations];
+
+      var newOrigRoomData = [...newOccupations[y]];
+      var room = newOrigRoomData[x];
+      newOrigRoomData[x] = undefined;
+      newOccupations[y] = newOrigRoomData;
+
+      var newDestRoomData = (newOccupations[targetY] === undefined) ? [] : [...newOccupations[targetY]];
+      newDestRoomData[x] = room;
+      newOccupations[targetY] = newDestRoomData;
+
+      return newOccupations;
+    } else {
+      return occupations;
+    }
+  }
+
+  function recalculateOccupations() {
+    var occupations = [];
+    for (var key in tiles) {
+      var tile = tiles[key];
+      var roomNumber = tile.roomNumber;
+      var row = occupations[roomNumber];
+      if (row === undefined) {
+        row = [];
+      }
+      var fromDate = new Date(Date.parse(tile.from.substr(0, 4) + '-' + tile.from.substr(4, 2) + '-' + tile.from.substr(6, 2)));
+      var x = Math.ceil((fromDate - date) / 86400000);
+      row[x] = tile;
+      occupations[roomNumber] = row;
+    }
+    return occupations;
+  }
+
+  function occupationsReducer(occupations, action) {
+    switch (action.type) {
+      case "move":
+        return handleMove(occupations, action.x, action.y, action.pageY);
+      case "dateChange":
+        return recalculateOccupations();
+      default:
+        break;
+    }
+  }
+
+  const [occupations, occupationsDispatch] = useReducer(occupationsReducer, recalculateOccupations());
+
+  function handleDateChange(event) {
+    var date = new Date(event.target.value);
+    setDate(date);
+    occupationsDispatch({type: "dateChange"});
+  }
+
   return(
     <div className="app">
-      <Header columns={columns}/>
-      <Table columns={columns} />
+      <Header date={date} onDateChange={handleDateChange} columns={columns} />
+      <Table date={date} rooms={rooms} tiles={tiles} occupations={occupations} occupationsDispatch={occupationsDispatch} columns={columns} />
     </div>
   );
 }
